Converted interface to abstract class to prevent code duplication, and removed unused public getters
